Fixed bug where nothing could be destroyed on hardware (Fixed issue 32).
This what caused by when I added the <= to the setArray function. Since it was assessing an extra argument, the DS and emulator interpreted it in different ways. The DS set the null argument to 0 (Making AIR a sword item, hence disabling block destroy), whereas I don't know what the emulator did.

**facepalm**